data_pipeline/pyproject.toml
[tool.poetry]
name = "sunbeam"
version = "0.1.0"
description = "UBC Solar's data pipeline, distribution, and storage cluster"
authors = ["Joshua Riefman <joshuariefman@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fusepy = "^3.0.1"
influxdb-client = "^1.45.0"
ubc-solar-data-tools = {git = "https://github.com/UBC-Solar/data_tools", rev = "main"}
prefect = "^2.20.2"
numpy = "^2.0.1"
pandas = "^2.2.2"
scipy = "^1.14.0"
toml = "^0.10.2"
pydantic = "^2.8.2"
tqdm = "^4.66.5"
prefect-github = "^0.2.7"
pymongo = "^4.8.0"

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"
